agents/requirements.txt
requests
yfinance
python-dotenv
google-generativeai